backend/requirements.txt
aioredis==1.3.1
amqp==5.0.6
aniso8601==7.0.0
anyio==3.6.2
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.6.0
django-countries==7.2.1
django-crequest==2018.5.11
django-filter==2.4.0
django-graphql-jwt==0.3.4
django-import-export==2.8.0
djangorestframework
django-seed==0.2.2
django-versatileimagefield==2.1
et-xmlfile==1.1.0
Faker==8.10.1
fcm-django==1.0.12
firebase-admin==5.4.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.1.0
MarkupPy==1.14
meilisearch==0.19.0
msgpack==1.0.2
numpy==1.23.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.0
Pillow==9.3.0
promise==2.3
prompt-toolkit==3.0.19
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.4.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-magic==0.4.24
pytimeparse==1.1.8
pytz==2021.1
PyYAML==6.0
redis==4.1.4
requests==2.26.0
requests-oauthlib==1.3.0
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.6.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.1
tablib==3.2.1
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.19.0
uvloop==0.16.0
vine==5.0.0
watchfiles==0.18.1
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
